Added JSON for test functionality.
